requirements.txt
Cython==0.29.23
dwave-ocean-sdk==4.3.0
dwave-networkx==0.8.10
h5py==2.10.0
lightfm==1.16
nltk==3.6.1
nose==1.3.7
numpy==1.21.5
pandas==1.2.4
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.6.2
tables
tqdm==4.59.0
